fix: catch errors and move utils to update
